Attempt to align NYC air quality data with proper timestamps
The 'realtime' NYC air quality data is actually delayed 2-4 hours. But Prometheus's textfile collector assumes that the timestamp of any datapoint is the mtime of the text file. The best way I could find to solve this problem was to hack the mtime of the text file. Prometheus is very opinionated and doesn't make this easy.