increase prometheus storage from 15d to 5y
this suggests that RAM is not usually a problem - mainly disk size is the bottleneck: https://www.robustperception.io/how-long-can-prometheus-retention-be/

The DB is < 3 MB for 15d. This suggests that for 5y, the DB size would be 365 * 5 / 15 * 3 = 365 MB, which I have plenty of space for.

If RAM is a problem, this blog post has some tips: https://www.robustperception.io/limiting-promql-resource-usage/

It discusses a limit on how many samples a PromQL query can have in memory at once: https://github.com/prometheus/prometheus/pull/4513